refactor(catalog): extract update-availability seam from tui
Move resolve_latest_digest and outdated_from_resolve out of
tui::update_check into a new src/catalog/update_availability.rs. The
command layer (grim status) needs the exact same "is a newer version
available" decision the TUI's background re-check uses, and it must
not depend on the TUI module tree to get it. Pure Two-Hats move: no
behavior change, no new tests. The outdated_from_resolve truth-table
unit tests move with it; resolve_latest_digest has no test of its own
today (it is only exercised transitively through UpdateChecker's
spawn_row_checks tests, which stay put since they test the TUI's
background-task machinery, not the pure function in isolation).

Refs: #43
